+++ src/char/int_storage.h
@@ -6,8 +6,8 @@ int inter_storage_init(void);
 void inter_storage_final();
 int inter_storage_save();
 int inter_guild_storage_save();
-int inter_storage_delete(unsigned long account_id);
-int inter_guild_storage_delete(unsigned long guild_id);
+int inter_storage_delete(uint32 account_id);
+int inter_guild_storage_delete(uint32 guild_id);
 
 int inter_storage_parse_frommap(int fd);
